HEAT

+++ 0x00-hello_world/6-size.c
@@ -13,7 +13,7 @@ int main(void)
 	float f;
 
 	printf("Size of a char: %lu byte (S)\n", (unsigned long)sizeof(d));
-	printf("Size of an int: %lu byte (S)\n", (unsigned long int)sizeof(a));
+	printf("Size of an int: %lu byte (S)\n", (unsigned int)sizeof(a));
 	printf("Size of a long int: %lu byte (S)\n", (unsigned long)sizeof(b));
 	printf("Size of a long long int: %lu byte (S)\n", (unsigned long)sizeof(c));
 	printf("Size of a float: %lu byte (S)\n", (unsigned long)sizeof(f));
